Many improvements.

+++ Sources/PlusAuth/RefreshAccessToken.swift
@@ -12,51 +12,62 @@ func refreshAccessToken(request: HTTPRequest, response: HTTPResponse) {
 				with: Data(bytes: bytes),
 				options: []
 				) as? [String: Any],
-			let refreshToken = json["refreshToken"] as? String
+			let refreshToken = json["refreshToken"] as? String,
+			let appId = json["appId"] as? String
 		else {
 			throw Err.request
 		}
 		
 		let storage = try PlusAuth.shared.storage()
-		
-		let accessToken = try Tokens.getAccessToken(bearer: request.header(.authorization))
-		
-		let (expiration, userId) = try storage.selectRefreshToken(token: refreshToken)
-		
-		let verified = try storage.selectUserVerified(byId: userId)
-		
-		guard verified else {
-			throw Err.unverified
+		try storage.startTransaction()
+		do {
+			let accessToken = try Tokens.getAccessToken(bearer: request.header(.authorization))
+			
+			let (expiration, userId, tokenAppId) = try storage.selectRefreshToken(token: refreshToken)
+			
+			guard tokenAppId == appId else {
+				throw Err.invalid
+			}
+			
+			let verified = try storage.selectUserVerified(byId: userId)
+			
+			guard verified else {
+				throw Err.unverified
+			}
+			
+			let now = Int64(Date().timeIntervalSince1970)
+			
+			guard now < expiration else {
+				try storage.deleteRefreshToken(token: refreshToken)
+				throw Err.expired
+			}
+			
+			let oldPayload = try Tokens.getAccessTokenPayload(accessToken: accessToken)
+			
+			guard
+				let sub = oldPayload["sub"] as? String,
+				sub == userId
+				else {
+					throw Err.invalid
+			}
+			
+			let newAccessToken = try Tokens.refreshAccessToken(oldAccessToken: accessToken)
+			
+			// TODO: do we need to regenerate new refresh token? Explore best practices
+			
+			let body = [
+				"accessToken": newAccessToken
+			]
+			try storage.commit()
+			response.setBody(bytes: body.toBytes())
+			response.status = .ok
+		} catch (let error) {
+			try storage.rollback()
+			throw error
 		}
-		
-		let now = Int64(Date().timeIntervalSince1970)
-		
-		guard now < expiration else {
-			try storage.deleteRefreshToken(token: refreshToken)
-			throw Err.expired
-		}
-		
-		let oldPayload = try Tokens.getAccessTokenPayload(accessToken: accessToken)
-		
-		guard
-			let sub = oldPayload["sub"] as? String,
-			sub == userId
-		else {
-			throw Err.invalid
-		}
-		
-		let newAccessToken = try Tokens.refreshAccessToken(oldAccessToken: accessToken)
-		
-		// TODO: do we need to regenerate new refresh token? Explore best practices
-		
-		let body = [
-			"accessToken": newAccessToken
-		]
-		
-		response.setBody(bytes: body.toBytes())
-		response.status = .ok
-
 	} catch Err.unverified {
+		response.status = .methodNotAllowed
+	} catch Err.invalid {
 		response.status = .methodNotAllowed
 	} catch Err.request {
 		response.status = .badRequest
